# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UCDT climate API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/alerts/read": {
      "post": {
        "tags": [
          "alerts"
        ],
        "summary": "Mark Read",
        "operationId": "mark_read",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkReadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/simulation": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Simulate",
        "description": "Live what-if against the latest weather/flood/aqi snapshots (what legacy runSimulation fetched).",
        "operationId": "simulate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest",
                "default": {
                  "cityId": "hcmc",
                  "scenario": {
                    "rainfallIncrease": 50.0,
                    "rainfallDurationHours": 3.0,
                    "addGreenCoverage": 0.0,
                    "trafficReduction": 0.0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResult"
                }
              }
            }
          },
          "503": {
            "description": "no snapshot yet"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/history/{hazard}": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "History",
        "description": "Snapshots of the last `hours`, oldest first.",
        "operationId": "history",
        "parameters": [
          {
            "name": "hazard",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hazard"
            }
          },
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 168,
              "minimum": 1,
              "default": 24,
              "title": "Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntry"
                  },
                  "title": "Response History"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FloodArea": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "riskLevel": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Risklevel"
          },
          "riskScore": {
            "type": "number",
            "title": "Riskscore"
          },
          "estimatedDepth": {
            "type": "number",
            "title": "Estimateddepth"
          },
          "geojson": {
            "$ref": "#/components/schemas/FloodGeoJSONFeature"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "name",
          "lat",
          "lng",
          "riskLevel",
          "riskScore",
          "estimatedDepth",
          "geojson"
        ],
        "title": "FloodArea"
      },
      "FloodGeoJSONFeature": {
        "properties": {
          "type": {
            "type": "string",
            "const": "Feature",
            "title": "Type"
          },
          "geometry": {
            "$ref": "#/components/schemas/PolygonGeometry"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "type",
          "geometry",
          "properties"
        ],
        "title": "FloodGeoJSONFeature"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryEntry": {
        "properties": {
          "computedAt": {
            "type": "string",
            "title": "Computedat"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result",
            "description": "that hazard's /latest shape as it was at computedAt"
          }
        },
        "type": "object",
        "required": [
          "computedAt",
          "result"
        ],
        "title": "HistoryEntry"
      },
      "MarkReadRequest": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "MarkReadRequest"
      },
      "MarkReadResponse": {
        "properties": {
          "marked": {
            "type": "integer",
            "title": "Marked",
            "description": "how many of the ids were unread and are now read"
          }
        },
        "type": "object",
        "required": [
          "marked"
        ],
        "title": "MarkReadResponse"
      },
...
      "PolygonGeometry": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "coordinates": {
            "items": {
              "items": {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Coordinates"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "type",
          "coordinates"
        ],
        "title": "PolygonGeometry"
      },
...
      "Scenario": {
        "properties": {
          "rainfallIncrease": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Rainfallincrease",
            "default": 50
          },
          "rainfallDurationHours": {
            "type": "number",
            "maximum": 72.0,
            "minimum": 0.0,
            "title": "Rainfalldurationhours",
            "default": 3
          },
          "addGreenCoverage": {
            "type": "number",
            "maximum": 100.0,
            "minimum": -100.0,
            "title": "Addgreencoverage",
            "description": "percentage points vs baseline",
            "default": 0
          },
          "trafficReduction": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Trafficreduction",
            "default": 0
          },
          "urbanDensity": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Urbandensity",
            "description": "omitted = baseline density"
          }
        },
        "type": "object",
        "title": "Scenario",
        "description": "Bounds = legacy validateScenario. Omitted/null levers take the legacy controller defaults."
      },
      "SimComparison": {
        "properties": {
          "before": {
            "$ref": "#/components/schemas/SimComparisonSide"
          },
          "after": {
            "$ref": "#/components/schemas/SimComparisonSide"
          }
        },
        "type": "object",
        "required": [
          "before",
          "after"
        ],
        "title": "SimComparison"
      },
      "SimComparisonSide": {
        "properties": {
          "riskScore": {
            "type": "number",
            "title": "Riskscore"
          },
          "affectedAreas": {
            "type": "integer",
            "title": "Affectedareas"
          }
        },
        "type": "object",
        "required": [
          "riskScore",
          "affectedAreas"
        ],
        "title": "SimComparisonSide"
      },
      "SimResults": {
        "properties": {
          "floodRiskDelta": {
            "type": "number",
            "title": "Floodriskdelta"
          },
          "newFloodAreas": {
            "items": {
              "$ref": "#/components/schemas/FloodArea"
            },
            "type": "array",
            "title": "Newfloodareas"
          },
          "tempDelta": {
            "type": "number",
            "title": "Tempdelta"
          },
          "aqiDelta": {
            "type": "number",
            "title": "Aqidelta"
          },
          "affectedBuildings": {
            "type": "integer",
            "title": "Affectedbuildings"
          },
          "affectedPopulation": {
            "type": "integer",
            "title": "Affectedpopulation"
          }
        },
        "type": "object",
        "required": [
          "floodRiskDelta",
          "newFloodAreas",
          "tempDelta",
          "aqiDelta",
          "affectedBuildings",
          "affectedPopulation"
        ],
        "title": "SimResults"
      },
      "SimulationRequest": {
        "properties": {
          "cityId": {
            "type": "string",
            "title": "Cityid",
            "default": "hcmc"
          },
          "scenario": {
            "$ref": "#/components/schemas/Scenario"
          }
        },
        "type": "object",
        "title": "SimulationRequest"
      },
      "SimulationResult": {
        "properties": {
          "simulationId": {
            "type": "string",
            "title": "Simulationid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "results": {
            "$ref": "#/components/schemas/SimResults"
          },
          "comparison": {
            "$ref": "#/components/schemas/SimComparison"
          }
        },
        "type": "object",
        "required": [
          "simulationId",
          "status",
          "results",
          "comparison"
        ],
        "title": "SimulationResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}